preliminary support for single data items
added plane, line, small circle as basic examples and slope as advanced. The basic API is just a QInputDialog for the attitude separated by spaces. Missing settings dialogs and the ability to change the attitude without deleting and reading. Need to start working on a better context menu with specific items. Maybe some sort of syntax like the rest of the data models?